+++ lib/features/stock_transfer_stocktake/controllers/stocktake_controller.dart
@@ -139,7 +139,7 @@ class StocktakeController extends ChangeNotifier with LoadState {
     notifyListeners();
   }
 
-  void setActualQuantity(StocktakeLine line, int? quantity) {
+  void setActualQuantity(StocktakeLine line, double? quantity) {
     line.actualQuantity = quantity;
     notifyListeners();
   }
